django/requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
pytz==2022.5
sqlparse==0.4.3
psycopg2-binary==2.9.4
parsedatetime==2.6
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
